Don't cache ENABLE_ variables
Summary:
QtCurve has a specialist detection of whether a variable is explicitly
set or not. If unset it probes for packages quietly. If explicitly set
to ON it will try to load the packages and error if they're not
existing.

By using option we were always caching the variable meaning compiling
twice would treat it like we ENABLE_QT5 was set.

Test Plan:
Ran on a machine without Qt5 twice, it continued gracefully saying Qt5 wasn't found
Set -DENABLE_QT5=ON it errored at the specific point.

Reviewers: yuyichao, rjvbb

Reviewed By: yuyichao

Differential Revision: https://phabricator.kde.org/D6341