reduce progressbar animation timer overhead
QtCurve ran a 20Hz timer to support animated and indeterminate progress
bars; this timer remained active at least as long as the widget remained
visible regardless if there was anything to animate.
This change starts the timer only when required and kills it as soon as
there is nothing left to animate.

Differential Revision: https://phabricator.kde.org/D7640
Committed from host : Portia.local