[qt5] Fix SpinMirror drawing in kalarm
The fix is basically copied from `QCommonStyle`. The issue appears to be
a shift in the coordinate causing the wrong edge to be exposed.
However, the whole coordinate calculation is too complex that I don't think
I know exactly what the coordinates are supposed to be.

Right-to-left rendering was tested in a hacy way by overriding `option`
in the `CC_SpinBox` rendering function.